Windows permission tests pass
