+++ rust/headless-client/src/ipc_service/windows.rs
@@ -2,8 +2,18 @@ use crate::CliCommon;
 use anyhow::{bail, Context as _, Result};
 use firezone_bin_shared::platform::DnsControlMethod;
 use futures::future::{self, Either};
-use std::{ffi::OsString, pin::pin, time::Duration};
+use std::{
+    ffi::{c_void, OsString},
+    mem::size_of,
+    pin::pin,
+    time::Duration,
+};
 use tokio::sync::mpsc;
+use windows::Win32::{
+    Foundation::{CloseHandle, HANDLE},
+    Security::{GetTokenInformation, TokenElevation, TOKEN_ELEVATION, TOKEN_QUERY},
+    System::Threading::{GetCurrentProcess, OpenProcessToken},
+};
 use windows_service::{
     service::{
         ServiceAccess, ServiceControl, ServiceControlAccept, ServiceErrorControl, ServiceExitCode,
@@ -15,6 +25,65 @@ use windows_service::{
 
 const SERVICE_NAME: &str = "firezone_client_ipc";
 const SERVICE_TYPE: ServiceType = ServiceType::OWN_PROCESS;
+
+/// Returns true if the IPC service can run properly
+pub(crate) fn elevation_check() -> Result<bool> {
+    let token = ProcessToken::our_process()?;
+    let elevated = token.is_elevated()?;
+    Ok(elevated)
+}
+
+// https://stackoverflow.com/questions/8046097/how-to-check-if-a-process-has-the-administrative-rights/8196291#8196291
+struct ProcessToken {
+    inner: HANDLE,
+}
+
+impl ProcessToken {
+    fn our_process() -> Result<Self> {
+        // SAFETY: Calling C APIs is unsafe
+        // `GetCurrentProcess` returns a pseudo-handle which does not need to be closed.
+        // Docs say nothing about thread safety. <https://learn.microsoft.com/en-us/windows/win32/api/processthreadsapi/nf-processthreadsapi-getcurrentprocess>
+        let our_proc = unsafe { GetCurrentProcess() };
+        let mut inner = HANDLE::default();
+        // SAFETY: We just created `inner`, and moving a `HANDLE` is safe.
+        // We assume that if `OpenProcessToken` fails, we don't need to close the `HANDLE`.
+        // Docs say nothing about threads or safety: <https://learn.microsoft.com/en-us/windows/win32/api/processthreadsapi/nf-processthreadsapi-openprocesstoken>
+        unsafe { OpenProcessToken(our_proc, TOKEN_QUERY, &mut inner) }
+            .context("`OpenProcessToken` failed")?;
+        Ok(Self { inner })
+    }
+
+    fn is_elevated(&self) -> Result<bool> {
+        let mut elevation = TOKEN_ELEVATION::default();
+        let token_elevation_sz = u32::try_from(size_of::<TOKEN_ELEVATION>())
+            .expect("`TOKEN_ELEVATION` size should fit into a u32");
+        let mut return_size = 0u32;
+        // SAFETY: Docs say nothing about threads or safety <https://learn.microsoft.com/en-us/windows/win32/api/securitybaseapi/nf-securitybaseapi-gettokeninformation>
+        // The type of `elevation` varies based on the 2nd parameter, but we hard-coded that.
+        // It should be fine.
+        unsafe {
+            GetTokenInformation(
+                self.inner,
+                TokenElevation,
+                Some(&mut elevation as *mut _ as *mut c_void),
+                token_elevation_sz,
+                &mut return_size as *mut _,
+            )
+        }?;
+        Ok(elevation.TokenIsElevated == 1)
+    }
+}
+
+impl Drop for ProcessToken {
+    fn drop(&mut self) {
+        // SAFETY: We got `inner` from `OpenProcessToken` and didn't mutate it after that.
+        // Closing a pseudo-handle is a harmless no-op, though this is a real handle.
+        // <https://learn.microsoft.com/en-us/windows/win32/api/processthreadsapi/nf-processthreadsapi-getcurrentprocess>
+        // > The pseudo handle need not be closed when it is no longer needed. Calling the CloseHandle function with a pseudo handle has no effect. If the pseudo handle is duplicated by DuplicateHandle, the duplicate handle must be closed.
+        unsafe { CloseHandle(self.inner) }.expect("`CloseHandle` should always succeed");
+        self.inner = HANDLE::default();
+    }
+}
 
 pub(crate) fn install_ipc_service() -> Result<()> {
     let manager_access = ServiceManagerAccess::CONNECT | ServiceManagerAccess::CREATE_SERVICE;
@@ -76,6 +145,9 @@ fn fallible_service_run(
     logging_handle: firezone_logging::file::Handle,
 ) -> Result<()> {
     tracing::info!(?arguments, "fallible_windows_service_run");
+    if !elevation_check()? {
+        bail!("IPC service failed its elevation check, try running as admin / root");
+    }
 
     let rt = tokio::runtime::Runtime::new()?;
     let (shutdown_tx, shutdown_rx) = mpsc::channel(1);
